change all buttons
