+++ app/routes/_layout-app/_mini-projetos/mini-projetos_.$slug_/_tabs/_overview/components/join-challenge-section.tsx
@@ -4,7 +4,7 @@ import LoadingButton from "~/components/features/form/loading-button";
 import type { Step } from "../../../build-steps.server";
 import DiscordButton from "~/components/features/auth/discord-button";
 import { BsDiscord } from "react-icons/bs";
-import { NewButton } from "~/components/ui/new-button";
+import { Button } from "~/components/ui/button";
 import { Card } from "~/components/ui/cards/card";
 import { NewInput } from "~/components/ui/new-input";
 import party from "party-js";
@@ -214,7 +214,7 @@ function StepForm({
   function getSecondaryButton() {
     return (
       step.secondaryButton && (
-        <NewButton
+        <Button
           type="submit"
           variant="outline"
           className=""
@@ -222,7 +222,7 @@ function StepForm({
           value={step.secondaryIntent}
         >
           {step.secondaryButton}
-        </NewButton>
+        </Button>
       )
     );
   }
